requirements.txt
django>=4.0
djangorestframework
gunicorn
django-cors-headers
psycopg2-binary
python_dotenv
